tg_bot_forwarder/requirements.txt
telethon
configparser
peewee
psycopg2-binary
sentry-sdk
python-telegram-bot
requests
